Add in algorithms and buttons js files
