+++ Scripts/algorithms.js
@@ -1,0 +1,62 @@
+import { globalState, updateSegmentElementsList } from './globalData.js';
+import htmlElements from './globalData.js';
+
+// Algorithm buttons
+htmlElements.algorithm1Button.addEventListener("click", () => {segment(1)});
+htmlElements.algorithm2Button.addEventListener("click", () => {segment(2)});
+htmlElements.algorithm3Button.addEventListener("click", () => {segment(3)});
+htmlElements.algorithm4Button.addEventListener("click", () => {segment(4)});
+
+// Import button
+htmlElements.importButton.addEventListener('click', async () => {
+    const filePaths = await window.api.openFile();
+    if (filePaths && filePaths.length > 0) {
+        // Display the file path
+        console.log('File path:', filePaths[0]); // This is available in Electron or environments with full file access
+        globalState.filePath = filePaths[0];
+        htmlElements.regions.clearRegions();
+        htmlElements.wavesurfer.load(filePaths[0]);
+    } else {
+        console.log('No file selected');
+    }
+});
+
+// runs the segmentation algorithm
+async function segment(algorithm) {
+    // const inputName = "C:\\Users\\sethb\\OneDrive - Worcester Polytechnic Institute (wpi.edu)\\gr-MQP-MLSongMap\\General\\Songs and Annotations\\Songs\\0043Carly Rae Jepsen  Call Me Maybe.wav"; // Example input data
+    const inputName = globalState.filePath;
+    globalState.clusters = determineVariability();
+    try {
+        console.log("Segmenting begin");
+
+        // Send a POST request to the Python server
+        const response = await fetch('http://127.0.0.1:5000/call-python', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({ song: inputName, algorithm : algorithm, clusters : globalState.clusters }),
+        });
+
+        console.log("Segmenting end");
+
+        // Parse the JSON response
+        const data = await response.json();
+
+        globalState.segmentData = data.map(row => {
+            return Object.fromEntries(row.map((value, index) => [globalState.headers[index], value]));
+        });
+
+        updateSegmentElementsList(globalState.segmentData, true)
+    } catch (error) {
+        console.error('Error:', error);
+    }
+}
+
+// Determines the variability to be used for an algorithm
+function determineVariability() {
+    const target = document.getElementById('variability-slider');
+    let num = parseInt(Number(target.value)/100) + 2
+    console.log(num)
+    return num
+}
