+++ page.html
@@ -60,7 +60,6 @@
                 <button class="btn" id="zoom-in">
                     <img src="resources/icons/zoom-in.svg" alt="Zoom In Button">
                 </button>
-                <p id="filePath">Selected file path will appear here.</p>
             </div>
         </div>
 
@@ -152,8 +151,9 @@
 
     <!-- <script src="scripts/algopy.js"></script> -->
     <script type="module" src="scripts/globalData.js"></script>
-    <script type="module" src="scripts/page.js"></script>
+    <script type="module" src="scripts/buttons.js"></script>
     <script type="module" src="scripts/editBoudaries.js"></script>
+    <script type="module" src="scripts/algorithms.js"></script>
 </body>
 
 </html>
